requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pip==23.3.1
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
PyJWT==1.7.1
pymongo==4.5.0
python-decouple==3.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart== 0.0.6
PyYAML==6.0.1
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.15.0
watchfiles==0.21.0
websockets==12.0